reducer refactoring
